+++ src/ResultSet.php
@@ -177,35 +177,7 @@ class ResultSet implements TabularDataReader, JsonSerializable
 
     public function select(string|int ...$columns): TabularDataReader
     {
-        $header = [];
-        $documentHeader = $this->getHeader();
-        $hasNoHeader = [] === $documentHeader;
-        foreach ($columns as $field) {
-            if (is_string($field)) {
-                if ($hasNoHeader) {
-                    throw new InvalidArgument(__METHOD__.' can only use named column if the tabular data has a non-empty header.');
-                }
-
-                $index = array_search($field, $this->header, true);
-                if (false === $index) {
-                    throw InvalidArgument::dueToInvalidColumnIndex($field, 'offset', __METHOD__);
-                }
-
-                $header[$index] = $field;
-                continue;
-            }
-
-            if (!$hasNoHeader && !array_key_exists($field, $documentHeader)) {
-                throw InvalidArgument::dueToInvalidColumnIndex($field, 'offset', __METHOD__);
-            }
-
-            $header[$field] = $documentHeader[$field] ?? $field;
-        }
-
-        /** @var array<int, string> $finalHeader */
-        $finalHeader = $hasNoHeader ? [] : $header;
-
-        return new self($this->combineHeader($header), $finalHeader);
+        return Statement::create()->select(...$columns)->process($this);
     }
 
     public function matching(string $expression): iterable